Detect/handle missing /boot device
Currently when the /boot device is missing, we just bail to a recovery
console. Now that we have a GUI that can configure the correct /boot
device, we should just detect the state of a missing /boot device,
prompt the user, and allow them to configure it correctly. This will be
useful as everyone uses pre-built Heads images without directly
configuring boot or USB devices.