Copy generated GPG key to USB
When you generate a GPG key on a GPG smart card, by default you don't
have an option to copy the GPG public key anywhere from within Heads.
This adds the option to copy a GPG public key you just generated to a
USB disk so you can then add it to the Heads keyring immediately.