mount_usb: fix handling of single USB device with > 1 partition
If a single USB device was attached and contained more than
one partition, instead of showing the menu, the block device
itself was automatically selected, and mounting failed.

Fix by making the menu selection conditional on the mount
device being set, and only set it in the single device /
single partition scenario.

Signed-off-by: Matt DeVillier <matt.devillier@puri.sm>